refactor: ErrorCode를 따로 패키지로 분리
