+++ src/main/java/joo/project/my3d/controller/ModelArticlesController.java
@@ -11,8 +11,8 @@ import joo.project.my3d.dto.response.ArticleDetailResponse;
 import joo.project.my3d.dto.response.ArticleFormResponse;
 import joo.project.my3d.dto.response.ArticlePreviewResponse;
 import joo.project.my3d.dto.security.BoardPrincipal;
-import joo.project.my3d.exception.ErrorCode;
 import joo.project.my3d.exception.FileException;
+import joo.project.my3d.exception.constant.ErrorCode;
 import joo.project.my3d.repository.ArticleLikeRepository;
 import joo.project.my3d.service.ArticleFileService;
 import joo.project.my3d.service.ArticleService;
