+++ src/main/java/joo/project/my3d/exception/ArticleLikeException.java
@@ -1,5 +1,6 @@
 package joo.project.my3d.exception;
 
+import joo.project.my3d.exception.constant.ErrorCode;
 import lombok.Getter;
 
 @Getter
