+++ src/main/java/joo/project/my3d/service/ArticleFileService.java
@@ -6,8 +6,8 @@ import joo.project.my3d.dto.ArticleFileDto;
 import joo.project.my3d.dto.DimensionDto;
 import joo.project.my3d.dto.DimensionOptionDto;
 import joo.project.my3d.dto.request.ArticleFormRequest;
-import joo.project.my3d.exception.ErrorCode;
 import joo.project.my3d.exception.FileException;
+import joo.project.my3d.exception.constant.ErrorCode;
 import joo.project.my3d.repository.ArticleFileRepository;
 import joo.project.my3d.service.aws.S3Service;
 import joo.project.my3d.utils.FileUtils;
