+++ src/main/java/joo/project/my3d/service/UserAccountService.java
@@ -7,8 +7,8 @@ import joo.project.my3d.dto.CompanyDto;
 import joo.project.my3d.dto.UserAccountDto;
 import joo.project.my3d.dto.properties.JwtProperties;
 import joo.project.my3d.dto.security.BoardPrincipal;
-import joo.project.my3d.exception.ErrorCode;
 import joo.project.my3d.exception.UserAccountException;
+import joo.project.my3d.exception.constant.ErrorCode;
 import joo.project.my3d.repository.AlarmRepository;
 import joo.project.my3d.repository.UserAccountRepository;
 import joo.project.my3d.utils.JwtTokenUtils;
