+++ src/test/java/joo/project/my3d/controller/ModelArticlesControllerTest.java
@@ -9,8 +9,8 @@ import joo.project.my3d.domain.constant.FormStatus;
 import joo.project.my3d.dto.*;
 import joo.project.my3d.dto.request.ArticleFormRequest;
 import joo.project.my3d.exception.ArticleException;
-import joo.project.my3d.exception.ErrorCode;
 import joo.project.my3d.exception.FileException;
+import joo.project.my3d.exception.constant.ErrorCode;
 import joo.project.my3d.fixture.Fixture;
 import joo.project.my3d.fixture.FixtureDto;
 import joo.project.my3d.repository.ArticleLikeRepository;
