+++ src/test/java/joo/project/my3d/service/UserAccountServiceTest.java
@@ -3,8 +3,8 @@ package joo.project.my3d.service;
 import joo.project.my3d.domain.UserAccount;
 import joo.project.my3d.domain.constant.UserRole;
 import joo.project.my3d.dto.UserAccountDto;
-import joo.project.my3d.exception.ErrorCode;
 import joo.project.my3d.exception.UserAccountException;
+import joo.project.my3d.exception.constant.ErrorCode;
 import joo.project.my3d.fixture.Fixture;
 import joo.project.my3d.fixture.FixtureDto;
 import joo.project.my3d.repository.AlarmRepository;
